# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server eshop server.",
        "title": "Eshop API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/",
    "paths": {
...
        "/admin_api/v1/wxapp_pages": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台自定义页面管理"
                ],
                "summary": "自定义页面列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台自定义页面管理"
                ],
                "summary": "添加自定义页面",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.WxAppPageParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
        "/admin_api/v1/wxapp_pages/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台自定义页面管理"
                ],
                "summary": "自定义页面详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "web_page id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台自定义页面管理"
                ],
                "summary": "更新自定义页面",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "web_page params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.WxAppPageParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台自定义页面管理"
                ],
                "summary": "删除自定义页面",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "web_page id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
...
        "/app_api/v1/addresses": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "地址"
                ],
                "summary": "获取用户的收获地址",
                "parameters": [
                    {
                        "description": "address pagination",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddressQueryParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "地址"
                ],
                "summary": "新增地址",
                "parameters": [
                    {
                        "description": "address params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddressParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
        "/app_api/v1/addresses/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "地址"
                ],
                "summary": "通过id获取地址",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "地址"
                ],
                "summary": "编辑地址",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "address attributes",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddressParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "地址"
                ],
                "summary": "删除地址",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
...
        "/app_api/v1/cities": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "地址"
                ],
                "summary": "根据省ID获取城市列表信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "province id",
                        "name": "province_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
...
        "/app_api/v1/regions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "地址"
                ],
                "summary": "根据市ID获取下级城市列表信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "city id",
                        "name": "city_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiHelpers.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "models.JSON": {
            "type": "array",
            "items": {
                "type": "integer"
            }
        },
...
        "v1.AddressParams": {
            "type": "object",
            "required": [
                "city",
                "detail",
                "is_default",
                "phone",
                "province",
                "receiver",
                "region"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "receiver": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "v1.AddressQueryParams": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "sort": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "v1.WxAppPageParams": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "page_data": {
                    "type": "object",
                    "$ref": "#/definitions/models.JSON"
                },
                "share_background_cover": {
                    "type": "string"
                },
                "share_cover": {
                    "type": "string"
                },
                "share_sentence": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}